Moved to SubtitleService
